refactor: P0~P4 아키텍처 리팩토링
P0: 환경변수 Joi 검증, Global Exception Filter, 이벤트 핸들러 try-catch, DiscordStrategy ConfigService 마이그레이션, 중복 DiscordModule.forRootAsync 제거
P1: JWT 인증 Guard, Global ValidationPipe, REST DTO 검증, 슬래시 커맨드 개별 파일 분리 및 discord.config 등록
P2: VoiceChannelService 서비스 분리 (VoiceSessionService, VoiceTempChannelService), VoiceDailyRepository raw SQL 캡슐화
P3: VoiceStateDto 커스텀 예외, Redis TTL 상수화, 로깅 레벨 정상화
P4: DiscordGateway LRU 캐시 eviction 개선
