+++ apps/api/package.json
@@ -38,6 +38,8 @@
     "@nestjs/platform-express": "^10.4.20",
     "@nestjs/swagger": "^7.4.2",
     "@nestjs/typeorm": "^10.0.2",
+    "class-transformer": "^0.5.1",
+    "class-validator": "^0.14.1",
     "date-fns-tz": "^3.2.0",
     "discord-player": "^7.1.0",
     "discord.js": "^14.16.3",
@@ -45,10 +47,12 @@
     "ffmpeg-static": "^5.2.0",
     "fluent-ffmpeg": "^2.1.3",
     "ioredis": "^5.8.2",
+    "joi": "^18.0.2",
     "jsonwebtoken": "^9.0.3",
     "opusscript": "^0.0.8",
     "passport": "^0.7.0",
     "passport-discord": "^0.1.4",
+    "passport-jwt": "^4.0.1",
     "pg": "^8.13.1",
     "play-dl": "^1.9.7",
     "prism-media": "^1.3.5",
@@ -66,6 +70,7 @@
     "@types/express": "^5.0.0",
     "@types/jest": "^29.5.2",
     "@types/node": "^20.3.1",
+    "@types/passport-jwt": "^4.0.1",
     "@types/supertest": "^6.0.0",
     "jest": "^29.5.0",
     "prettier": "^3.0.0",
