+++ apps/api/src/auth/discord.strategy.ts
@@ -1,13 +1,15 @@
-// auth/discord.strategy.ts
+import { Injectable } from '@nestjs/common';
+import { ConfigService } from '@nestjs/config';
 import { PassportStrategy } from '@nestjs/passport';
 import { Strategy } from 'passport-discord';
 
+@Injectable()
 export class DiscordStrategy extends PassportStrategy(Strategy, 'discord') {
-  constructor() {
+  constructor(configService: ConfigService) {
     super({
-      clientID: process.env.DISCORD_CLIENT_ID,
-      clientSecret: process.env.DISCORD_CLIENT_SECRET,
-      callbackURL: process.env.DISCORD_CALLBACK_URL,
+      clientID: configService.get<string>('DISCORD_CLIENT_ID'),
+      clientSecret: configService.get<string>('DISCORD_CLIENT_SECRET'),
+      callbackURL: configService.get<string>('DISCORD_CALLBACK_URL'),
       scope: ['identify', 'guilds'],
     });
   }
