+++ apps/api/src/channel/voice/voice-channel.module.ts
@@ -2,22 +2,23 @@ import { DiscordModule } from '@discord-nestjs/core';
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 
-import { DiscordConfig } from '../../config/discord.config';
 import { MemberModule } from '../../member/member.module';
 import { ChannelModule } from '../channel.module';
 import { VoiceChannelPolicy } from './application/voice-channel.policy';
 import { VoiceChannelService } from './application/voice-channel.service';
 import { VoiceChannelHistoryService } from './application/voice-channel-history.service';
 import { VoiceDailyFlushService } from './application/voice-daily-flush-service';
+import { VoiceSessionService } from './application/voice-session.service';
+import { VoiceTempChannelService } from './application/voice-temp-channel.service';
 import { VoiceChannelHistory } from './domain/voice-channel-history.entity';
 import { VoiceDailyEntity } from './domain/voice-daily.entity';
 import { DiscordVoiceGateway } from './infrastructure/discord-voice.gateway';
 import { RedisTempChannelStore } from './infrastructure/redis-temp-channel-store';
+import { VoiceDailyRepository } from './infrastructure/voice-daily.repository';
 import { VoiceRedisRepository } from './infrastructure/voice-redis.repository';
 
 @Module({
   imports: [
-    DiscordModule.forRootAsync(DiscordConfig),
     DiscordModule.forFeature(),
     TypeOrmModule.forFeature([VoiceChannelHistory, VoiceDailyEntity]),
     MemberModule,
@@ -25,6 +26,8 @@ import { VoiceRedisRepository } from './infrastructure/voice-redis.repository';
   ],
   providers: [
     VoiceChannelService,
+    VoiceSessionService,
+    VoiceTempChannelService,
     VoiceChannelHistoryService,
     VoiceChannelPolicy,
     DiscordVoiceGateway,
@@ -33,6 +36,7 @@ import { VoiceRedisRepository } from './infrastructure/voice-redis.repository';
       useClass: RedisTempChannelStore,
     },
     VoiceRedisRepository,
+    VoiceDailyRepository,
     VoiceDailyFlushService,
   ],
   exports: [VoiceChannelService, TypeOrmModule],
