+++ apps/api/src/config/base.config.ts
@@ -1,4 +1,10 @@
+import { envValidationSchema } from './env.validation';
+
 export const BaseConfig = {
   isGlobal: true,
   envFilePath: `../../.env`,
+  validationSchema: envValidationSchema,
+  validationOptions: {
+    abortEarly: false,
+  },
 };
