+++ apps/api/src/config/env.validation.ts
@@ -1,0 +1,33 @@
+import * as Joi from 'joi';
+
+export const envValidationSchema = Joi.object({
+  // Database
+  DATABASE_HOST: Joi.string().required(),
+  DATABASE_PORT: Joi.number().port().default(5432),
+  DATABASE_USER: Joi.string().required(),
+  DATABASE_PASSWORD: Joi.string().required(),
+  DATABASE_NAME: Joi.string().required(),
+
+  // Discord
+  DISCORD_API_TOKEN: Joi.string().required(),
+  DISCORD_CLIENT_ID: Joi.string().required(),
+  DISCORD_CLIENT_SECRET: Joi.string().optional(),
+  DISCORD_CALLBACK_URL: Joi.string().optional(),
+  DISCORD_COMMAND_PREFIX: Joi.string().default('!'),
+
+  // Redis
+  REDIS_HOST: Joi.string().required(),
+  REDIS_PORT: Joi.number().port().default(6379),
+  REDIS_PASSWORD: Joi.string().allow('').optional(),
+
+  // Gemini
+  GEMINI_API_KEY: Joi.string().required(),
+
+  // JWT
+  JWT_SECRET: Joi.string().required(),
+
+  // App
+  NODE_ENV: Joi.string().valid('development', 'production', 'test').default('development'),
+  PORT: Joi.number().port().default(3000),
+  WEB_URL: Joi.string().default('http://localhost:4000'),
+});
