--- apps/api/src/config/server.config.ts
@@ -1,4 +1,0 @@
-
-export const ServerConfig = {
-  SERVER_PORT: parseInt(process.env.PORT, 10) || 3000,
-};
