+++ apps/api/src/main.ts
@@ -1,14 +1,30 @@
-import { Logger } from '@nestjs/common';
+import { Logger, ValidationPipe } from '@nestjs/common';
+import { ConfigService } from '@nestjs/config';
 import { NestFactory } from '@nestjs/core';
 
 import { AppModule } from './app.module';
-import { ServerConfig } from './config/server.config';
+import { AllExceptionsFilter } from './common/filters/all-exceptions.filter';
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule);
 
-  await app.listen(ServerConfig.SERVER_PORT ?? 3000, async () => {
-    Logger.log(`Server listening on port ${ServerConfig.SERVER_PORT}`);
+  app.useGlobalFilters(new AllExceptionsFilter());
+  app.useGlobalPipes(
+    new ValidationPipe({
+      whitelist: true,
+      transform: true,
+    }),
+  );
+
+  const configService = app.get(ConfigService);
+  const port = configService.get<number>('PORT', 3000);
+
+  await app.listen(port, () => {
+    Logger.log(`Server listening on port ${port}`);
   });
 }
-bootstrap();
+
+bootstrap().catch((err) => {
+  Logger.error('Failed to start application', err);
+  process.exit(1);
+});
